Clean up logic and ready for test
