+++ index.js
@@ -3,7 +3,8 @@ require("dotenv").config();
 const connectToDatabase = require("./config/db");
 const multer = require('multer');
 const ReturnProcess = require('./models/ReturnProcess');
-const uploadImage = require('./middleware/uploadImage');
+const uploadImage = require('./middleware/upload');
+const path = require('path');
 const authRouter = require("./routes/authRoute");
 const orderRouter = require("./routes/ordersRoute");
 const paymentRouter = require("./routes/paymentRoute");
