--- middleware/uploadImage.js
@@ -1,15 +1,0 @@
-const path = require('path');
-const multer = require('multer');
-
-// Storage options are here.
-const storage = multer.diskStorage({
-    destination: function(req, file, cb) {
-        cb(null, 'uploads/');
-    },
-    filename: function(req, file, cb) {
-        const ext = path.extname(file.originalname);
-        cb(null, Date.now() + ext);
-    }
-});
-
-const upload = multer({ storage: storage });
