+++ package.json
@@ -16,8 +16,10 @@
     "dotenv": "^16.3.2",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
+    "mongodb": "^6.3.0",
     "mongoose": "^7.3.0",
-    "multer": "^1.4.5-lts.1",
+    "multer": "^1.4.2",
+    "multer-gridfs-storage": "^5.0.2",
     "nodemailer": "^6.9.3",
     "nodemon": "^3.0.3",
     "stripe": "^14.16.0"
